+++ platform/evaluatorResponse.go
@@ -1,6 +1,10 @@
 package platform
 
-import "github.com/robbyt/go-polyscript/platform/data"
+import (
+	"time"
+
+	"github.com/robbyt/go-polyscript/platform/data"
+)
 
 // EvaluatorResponse is based on the risor object.Object interface, but with some features removed
 type EvaluatorResponse interface {
@@ -13,9 +17,14 @@ type EvaluatorResponse interface {
 	// Interface converts the given object to a native Go value.
 	Interface() any
 
-	// GetScriptExeID returns the ID of the script that generated the object.
-	GetScriptExeID() string
+	// ScriptExeID returns the ID of the script that generated the object.
+	ScriptExeID() string
 
-	// GetExecTime returns the time it took to execute the script
-	GetExecTime() string
+	// ExecTime returns the time it took to execute the script.
+	ExecTime() time.Duration
+
+	// AsMap returns the result as a map[string]any. Returns an error of the
+	// form "AsMap: expected map[string]any, got <T>" when the underlying value
+	// has a different shape.
+	AsMap() (map[string]any, error)
 }